No need to call ShutdownProtobufLibrary
I had added it to make sure leak checkers would see
no leaks, because of the comment at the bottom of
https://protobuf.dev/getting-started/cpptutorial/#writing-a-message.
However, Valgrind says there are no leaks even
without it.

Note that the code removed was wrong anyways:
in from_file, the function was only called in
case of errors. Thanks Martin!